Fix dataset delete
Deleting a dataset using psql as the database failed because a soundevent is
attached to a recording. When deleting a dataset, all recordings in this
dataset are deleted, resulting in a psql foreign key violation, because each
sound event needs a recording.
In this fix, we first delete all soundevents associated to all recordings in
the dataset and then delete it.