Fix audio playback bug
The audio stream is sent in chunkes, where the response code from the
server is 206 PARTIAL. This is correct. However, when the last chunk is sent,
Chrome stopped playback with the error message
"PIPELINE_ERROR_READ: FFmpegDemuxer: data source error".
To fix this issue, the response code of the very last chunk should be 200 OK.