Introduce better home button handling
In the annotation view the home button in the spectrogram
controlls always caused the spectrogram to jump to the beginning.
This behaviour is now changed.
When the view was dragged only, the behaviour stays the same, i.e.,
the home button will bring you home.
But of the user zoomed in the home button will restore the last drag
position.