Add new zoom in and out buttons
The spectrogram view now behaves somewhat differently.
By default, the drag command is selected, as usualy.
But there are now three zoom modes: first, zoom to selection.
This is done as usual. Second, zoom in and third, zoom out.
The last two modes allow to strecht or compress the view.
Finally, there is no back to home button anymore. The back button
now restores the original zoom while keeping the start time on
the x-axis.