Use webP and jpeg spectrograms
As of now, whombat used PNGs as the spectrogram image format.
However, it took up to three seconds to compute those.
Interestingly, not the computation of the stft was the long running
task but writing the pillow image object to the result buffer.
We now use webP as the standard image format as encoding is way
faster. However, webP has the problem of only allowing 2^14-1
pixels per axis. If we have a spectrogram that is larger we
use jpeg as the fallback.