Fix resample bug
When resampling was activated in the spectrogram settings,
the viewpower was updated as well. That let to first showing
the resampled spectrogram and second to a wrongly zoomed viewport.
The fix is two-fold. First, the bounds parameter from the parent
component prevented re-rendering the y-axis. Second, the window
bounds calculation was not reactive so that changes were not
triggering a re-render.